pyproject.toml
[tool.poetry]
name = "jwt-auth-example"
version = "0.1.0"
description = ""
authors = ["Sergey Yaksanov <iaksanov2044@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
pre-commit = "^3.7.0"
djangorestframework-simplejwt = "^5.3.1"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
